Added the "Iulius" theme. This is a lightweight, practical, functional theme created by Nathan Ziarek. The wide space for menu titles, high contrast text and the simple wire frames offers a good alternative to the more graphic intensive themes and may be more appealing to the Sgt. Joe Friday, Frank Lloyd Wright, or even Adrian Monk mind sets.
I would be reluctant to have another theme to maintain, however, the
ui.xml file is identical to G.A.N.T.'s except for the color of the
"record_font". There is no reason these shouldn't be able to stay in
sync so diffs for one can be applied to the other and there will be
virtually no maintenance. In addition to offering users a clear
alternative, I believe this is also a good example of the minimum
requirements theme designers need in order to create an effective
theme.


git-svn-id: http://svn.mythtv.org/svn/trunk@4955 7dbf422c-18fa-0310-86e9-fd20926502f2