Please, please read this commit.  It's important for making what follows work and it'll make me feel like people like me.
Still here?

OK.  Surprise!  Commit Arclight to myththemes.  *Must* be used with trunk r23269 or later, and must isntall the included fonts systemwide, or it will look totally wrong.  Not using this version or later?  Didn't install the fonts?  Don't complain to me!

Things which aren't completely finished but will be in the near future:
--------------
MythWeather
MythArchive
A number of menu button watermarks
A bit of cleanup/reformatting in the program guide(s)
A few bits of mythtv-setup

Things which are unlikely ever to get done:
--------------
MythZoneminder

Everything else is pretty much done.  Some funkiness with the program guide occasionally due to bizarre scaling differences, which should hopefully be fixed in a change to how we do our rounding/scaling in the near future.  Seems to work properly at 1366x768 and 1280x720, but the guide can look a little off at 1920x1080.  May also get fixed as the guide widget is improved in the future.

What do you need to know to make the theme work right?  The theme includes two fonts, each redistributable, but each with their own license.  The theme will look *very, very wrong* if you don't install the fonts.  See the LICENSES file for more information if you are curious about them.  Both .otf files much be installed systemwide for the theme to work right.  For Ubuntu, that's something like:

sudo cp *.otf /usr/share/fonts/
sudo fc-cache -f

Oh, and the theme itself is GPL.  But not because you whined.



git-svn-id: http://svn.mythtv.org/svn/trunk@23275 7dbf422c-18fa-0310-86e9-fd20926502f2